opcion 1 revisa toda la oracion y opcion 5 puede tomar float

+++ include/math_utils.h
@@ -6,4 +6,4 @@ int Sumatoria(int *array, int length);
 int Promedio(int *array, int length);
 
 //Devuelve resultado de la funcion f(x) = 5x*x+1/x
-float Func1(int x);
+float Func1(float x);
