+++ src/math_utils.c
@@ -23,6 +23,6 @@ int Promedio(int *array, int length){
 }
 
 //Devuelve resultado de la funcion f(x) = 5x*x+1/x
-float Func1(int x){
+float Func1(float x){
     return (5 * x * x + (1.0/x));
 }
